LPS-56245 Move WebDAV tests to document library test module
