+++ modules/apps/document-library/document-library-test/test/integration/com/liferay/document/library/dependencies/OSX_Test.docx
