+++ modules/apps/document-library/document-library-test/test/integration/com/liferay/document/library/webdav/test/WebDAVEnvironmentConfigTestCallback.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.webdav;
+package com.liferay.document.library.webdav.test;
 
 import com.liferay.portal.kernel.util.Tuple;
 import com.liferay.portal.kernel.webdav.methods.Method;
